requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.10.1
requests==2.27.1
python-telegram-bot==13.11
gunicorn==20.1.0